+++ tests/routers/test_app_v1.py
@@ -18,6 +18,7 @@ limitations under the License.
 
 # Imports
 from server import app
+
 from fastapi.testclient import TestClient
 
 # Initializing the TestClient