+++ tests/routers/v1/test_entertainment.py
@@ -17,9 +17,10 @@ limitations under the License.
 """
 
 # Imports
+from server import app
+
 import respx
 import httpx
-from server import app
 from fastapi.testclient import TestClient
 
 # Initializing the TestClient